create base_field and add a field directory
